backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

# ML stack (Python 3.12 compatible)
torch==2.2.2
torchvision==0.17.2
sentence-transformers==2.6.1
transformers==4.38.2
numpy==1.26.4

# Backend deps - UPDATED VERSIONS
supabase==2.9.0
httpx==0.27.0
pydantic==2.5.0
python-dotenv==1.0.0groq==0.11.0
